Add human-set, clamped tool-call budget overrides (chat + coder loops)
Running a PR-sized task in the app hit a hardcoded tool-call limit that
forced a "send a follow-up to continue" restart, re-exploring from scratch
and wasting tokens. This adds two override knobs — both HUMAN-set and
clamped, so the autonomous loop can never raise its own limit (the
self-modification guard is preserved):

- Per-project Playground chat tool budget: new GET/POST /v1/chat-limits
  (chatlimits.go), default 6 rounds / 24 calls, compile-time ceiling
  24 / 96. effectiveChatLimits computes default -> stored(clamped) ->
  per-request header (LOWER-only) -> floor 1, mirroring BridgeMaxHops but
  comparing in float space to avoid its 1e300 float->int underflow (fixed
  here too as a companion). Persisted in new project_settings columns via a
  column-specific upsert that can't clobber auto_pr.
- Per-run coder tool budget: RunConfig.MaxToolCalls (default 40, clamp
  1..200), frozen at creation like MaxIterations; runCoder honors it with a
  legacy fallback to Engine.MaxToolCalls for pre-migration rows (new
  runs.max_tool_calls column). Surfaced in the pre-flight, run view, and the
  New-run modal.
- "Start over" mitigation: the cap now prompts for a Progress-Notes
  carry-forward and returns an additive l00prite_chat_tools field the
  dashboard renders as a hint with a one-click "Raise tool budget" button.
- Dashboard: Playground "Tool budget" modal, cap-hit hint, New-run
  "Max tool calls per unit" field, pre-flight Budgets line.
- Test-only: mock.go gains a /chattoolloop directive (the chat-tool analogue
  of /bridgeloop) to drive the round cap end to end.

propose_run's schema deliberately gains no budget field, so a chat-drafted
run always gets the server default. The EXECUTE gate is untouched. Defaults
are unchanged, so a gateway/run with no override behaves as before.

Verified: go test -race ./... green (new regression tests for the clamp,
effectiveChatLimits edge cases incl. 1e300, the BridgeMaxHops fix, migration,
the coder budget honored above 40, and the endpoint round-trip/coexistence);
gofmt/vet clean; validator 519 PASS / 0 FAIL; two review-gated files
untouched; live-verified against the real binary + Playwright.

+++ cli-os/internal/gateway/bridge.go
@@ -103,7 +103,13 @@ func BridgeMaxHops(h http.Header, cfg config.Config) int {
 	if err != nil || n < 0 || math.IsInf(n, 0) || math.IsNaN(n) {
 		return base
 	}
-	if int(n) < base {
+	// Compare in FLOAT space before converting to int. A finite-but-huge value like "1e300" passes
+	// the Inf/NaN guard, yet int(1e300) is implementation-dependent (min-int on amd64) -- so the old
+	// `int(n) < base` treated it as a request to LOWER the cap to a negative, making maxTurns
+	// negative and the bridge return a null 200. `n < float64(base)` cannot: 1e300 is never < base,
+	// so the header is simply ignored (the intended "may only lower" semantics). Matches
+	// effectiveChatLimits (chatlimits.go), which was written correctly from the start.
+	if n < float64(base) {
 		return int(n)
 	}
 	return base
